Правки по комментариям
