+++ ImageFeed/Constants.swift
@@ -7,8 +7,11 @@
 
 import Foundation
 
-let accessKey = "JQJgJax-Jbr97jBjYuQLDZgaP5DX-b5Ckx1JA1pe09E"
-let secretKey = "YL93TX8vKWwFQ6iZ9baPSxOEW_BGS_A8TUhJHLUlGVo"
-let redirectURI = "urn:ietf:wg:oauth:2.0:oob"
-let accessScope = "public+read_user+write_likes"
-let defaultBaseURL = URL(string: "https://api.unsplash.com")!
+struct Constants {
+    static let accessKey = "JQJgJax-Jbr97jBjYuQLDZgaP5DX-b5Ckx1JA1pe09E"
+    static let secretKey = "YL93TX8vKWwFQ6iZ9baPSxOEW_BGS_A8TUhJHLUlGVo"
+    static let redirectURI = "urn:ietf:wg:oauth:2.0:oob"
+    static let accessScope = "public+read_user+write_likes"
+    static let defaultBaseURL = URL(string: "https://api.unsplash.com")!
+}
+
