+++ ImageFeed/ServiceAccessLayer/OAuth2/OAuthTokenResponseBody.swift
@@ -1,0 +1,22 @@
+//
+//  OAuthTokenResponseBody.swift
+//  ImageFeed
+//
+//  Created by Глеб Хамин on 16.01.2024.
+//
+
+import Foundation
+
+struct OAuthTokenResponseBody: Decodable {
+        let accessToken: String
+        let tokenType: String
+        let scope: String
+        let createdAt: Int
+    
+        enum CodingKeys: String, CodingKey {
+            case accessToken = "access_token"
+            case tokenType = "token_type"
+            case scope
+            case createdAt = "created_at"
+        }
+}
